Move scan context provider into shared components directory
